Modified all iDistance sensors to return float.

+++ include/interfaces/iDistance.hpp
@@ -36,7 +36,7 @@ namespace upm
   {
   public:
     virtual ~iDistance() {}
-    virtual int getDistance() = 0;
+    virtual float getDistance() = 0;
     /**
       * Convert distance value from Cm(default) to one
       * of the following:
