+++ src/urm37/urm37.cxx
@@ -67,7 +67,7 @@ float URM37::getDistance(int degrees)
   return (distance);
 }
 
-int URM37::getDistance()
+float URM37::getDistance()
 {
     /* TODO: compilation issue for swig. switched original method to not use default zero parameter. */
     return getDistance(0);
